requirements.txt
click
jinja2
psycopg2-binary